+++ crates/config/src/workspace/mod.rs
@@ -8,9 +8,7 @@ use crate::constants;
 use crate::errors::map_figment_error_to_validation_errors;
 use crate::providers::url::Url;
 use crate::types::{FileGlob, FilePath};
-use crate::validators::{
-    default_bool_true, validate_child_relative_path, validate_extends_url, validate_id,
-};
+use crate::validators::{validate_child_relative_path, validate_extends, validate_id};
 use figment::value::{Dict, Map};
 use figment::{
     providers::{Format, Serialized, Yaml},
@@ -31,12 +29,6 @@ use validator::{Validate, ValidationError, ValidationErrors};
 pub use vcs::{VcsConfig, VcsManager};
 
 type ProjectsMap = HashMap<String, FilePath>;
-
-fn validate_extends(extends: &str) -> Result<(), ValidationError> {
-    validate_extends_url("extends", extends)?;
-
-    Ok(())
-}
 
 // Validate the `projects` field is a map of valid file system paths
 // that are relative from the workspace root. Will fail on absolute
@@ -61,7 +53,6 @@ fn validate_projects(projects: &ProjectsMap) -> Result<(), ValidationError> {
 #[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize, Validate)]
 #[serde(rename_all = "camelCase")]
 pub struct ActionRunnerConfig {
-    #[serde(default = "default_bool_true")]
     pub inherit_colors_for_piped_tasks: bool,
 }
 
@@ -77,28 +68,23 @@ impl Default for ActionRunnerConfig {
 #[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize, Validate)]
 #[serde(rename_all = "camelCase")]
 pub struct WorkspaceConfig {
-    #[serde(default)]
     #[validate]
     pub action_runner: ActionRunnerConfig,
 
     #[validate(custom = "validate_extends")]
     pub extends: Option<String>,
 
-    #[serde(default)]
     #[validate]
     pub node: NodeConfig,
 
-    #[serde(default)]
     #[serde(deserialize_with = "deserialize_projects")]
     #[schemars(schema_with = "make_projects_schema")]
     #[validate(custom = "validate_projects")]
     pub projects: ProjectsMap,
 
-    #[serde(default)]
     #[validate]
     pub typescript: TypeScriptConfig,
 
-    #[serde(default)]
     #[validate]
     pub vcs: VcsConfig,
 
@@ -117,7 +103,7 @@ impl Provider for WorkspaceConfig {
     }
 
     fn data(&self) -> Result<Map<Profile, Dict>, figment::Error> {
-        Serialized::defaults(WorkspaceConfig::default()).data()
+        Serialized::defaults(self).data()
     }
 
     fn profile(&self) -> Option<Profile> {
@@ -128,19 +114,24 @@ impl Provider for WorkspaceConfig {
 impl WorkspaceConfig {
     pub fn load(path: PathBuf) -> Result<WorkspaceConfig, ValidationErrors> {
         let mut config = WorkspaceConfig::load_config(
-            Figment::from(Serialized::defaults(WorkspaceConfig::default()))
-                .merge(Yaml::file(&path)),
+            Figment::from(WorkspaceConfig::default()).merge(Yaml::file(&path)),
         )?;
 
         // This is janky, but figment does not support any kind of extends mechanism,
         // and figment providers do not have access to the current config dataset,
         // so we need to double-load this config and extract in the correct order!
         if let Some(extends) = config.extends {
-            config = WorkspaceConfig::load_config(
-                Figment::from(Serialized::defaults(WorkspaceConfig::default()))
-                    .merge(Url::from(extends))
-                    .merge(Yaml::file(&path)),
-            )?;
+            let mut figment = Figment::from(WorkspaceConfig::default());
+
+            if extends.starts_with("http") {
+                figment = figment.merge(Url::from(extends));
+            } else {
+                figment = figment.merge(Yaml::file(path.parent().unwrap().join(extends)));
+            };
+
+            figment = figment.merge(Yaml::file(&path));
+
+            config = WorkspaceConfig::load_config(figment)?;
         }
 
         // Versions from env vars should take precedence
@@ -259,10 +250,10 @@ fn make_projects_schema(_gen: &mut SchemaGenerator) -> Schema {
 mod tests {
     use super::*;
     use crate::errors::tests::handled_jailed_error;
-    use std::env;
+    use std::path::Path;
 
-    fn load_jailed_config() -> Result<WorkspaceConfig, figment::Error> {
-        match WorkspaceConfig::load(PathBuf::from(constants::CONFIG_WORKSPACE_FILENAME)) {
+    fn load_jailed_config(root: &Path) -> Result<WorkspaceConfig, figment::Error> {
+        match WorkspaceConfig::load(root.join(constants::CONFIG_WORKSPACE_FILENAME)) {
             Ok(cfg) => Ok(cfg),
             Err(errors) => Err(handled_jailed_error(&errors)),
         }
@@ -273,7 +264,7 @@ mod tests {
         figment::Jail::expect_with(|jail| {
             jail.create_file(constants::CONFIG_WORKSPACE_FILENAME, "projects: {}")?;
 
-            let config = load_jailed_config()?;
+            let config = load_jailed_config(jail.directory())?;
 
             assert_eq!(
                 config,
@@ -294,6 +285,7 @@ mod tests {
 
     mod extends {
         use super::*;
+        use std::fs;
 
         #[test]
         #[should_panic(
@@ -303,22 +295,24 @@ mod tests {
             figment::Jail::expect_with(|jail| {
                 jail.create_file(super::constants::CONFIG_WORKSPACE_FILENAME, "extends: 123")?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
         }
 
         #[test]
-        #[should_panic(expected = "Invalid field <id>extends</id>: Must be a valid URL.")]
-        fn not_a_url() {
+        #[should_panic(
+            expected = "Invalid field <id>extends</id>: Must be a valid URL or relative file path (starts with ./)."
+        )]
+        fn not_a_url_or_file() {
             figment::Jail::expect_with(|jail| {
                 jail.create_file(
                     super::constants::CONFIG_WORKSPACE_FILENAME,
                     "extends: random value",
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -333,16 +327,14 @@ mod tests {
                     "extends: http://domain.com",
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
         }
 
         #[test]
-        #[should_panic(
-            expected = "Invalid field <id>extends</id>: Must be a YAML (.yml) document."
-        )]
+        #[should_panic(expected = "Invalid field <id>extends</id>: Must be a YAML document.")]
         fn not_a_yaml_url() {
             figment::Jail::expect_with(|jail| {
                 jail.create_file(
@@ -350,7 +342,65 @@ mod tests {
                     "extends: https://domain.com/file.txt",
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
+
+                Ok(())
+            });
+        }
+
+        #[test]
+        #[should_panic(expected = "Invalid field <id>extends</id>: Must be a YAML document.")]
+        fn not_a_yaml_file() {
+            figment::Jail::expect_with(|jail| {
+                fs::create_dir_all(jail.directory().join("shared")).unwrap();
+
+                jail.create_file("shared/file.txt", "")?;
+
+                jail.create_file(
+                    super::constants::CONFIG_WORKSPACE_FILENAME,
+                    "extends: ./shared/file.txt",
+                )?;
+
+                super::load_jailed_config(jail.directory())?;
+
+                Ok(())
+            });
+        }
+
+        #[test]
+        fn loads_from_file() {
+            figment::Jail::expect_with(|jail| {
+                fs::create_dir_all(jail.directory().join("shared")).unwrap();
+
+                jail.create_file(
+                    format!("shared/{}", super::constants::CONFIG_WORKSPACE_FILENAME),
+                    include_str!("../../../../tests/fixtures/config-extends/.moon/workspace.yml"),
+                )?;
+
+                jail.create_file(
+                    super::constants::CONFIG_WORKSPACE_FILENAME,
+                    r#"
+extends: ./shared/workspace.yml
+
+node:
+    version: '18.0.0'
+    npm:
+        version: '8.0.0'
+"#,
+                )?;
+
+                let config: WorkspaceConfig = super::load_jailed_config(jail.directory())?;
+
+                assert_eq!(config.extends, Some("./shared/workspace.yml".to_owned()));
+
+                // Inherits from extended file
+                assert!(!config.node.add_engines_constraint);
+                assert!(!config.typescript.sync_project_references);
+                assert_eq!(config.vcs.manager, VcsManager::Svn);
+
+                // Ensure we can override the extended config
+                assert_eq!(config.node.version, "18.0.0".to_owned());
+                assert_eq!(config.node.npm.version, "8.0.0".to_owned());
 
                 Ok(())
             });
@@ -371,7 +421,7 @@ node:
 "#,
                 )?;
 
-                let config: WorkspaceConfig = super::load_jailed_config()?;
+                let config: WorkspaceConfig = super::load_jailed_config(jail.directory())?;
 
                 assert_eq!(
                     config.extends,
@@ -406,7 +456,7 @@ node:
     packageManager: yarn"#,
                 )?;
 
-                let config = super::load_jailed_config()?;
+                let config = super::load_jailed_config(jail.directory())?;
 
                 assert_eq!(
                     config,
@@ -436,7 +486,7 @@ node:
             figment::Jail::expect_with(|jail| {
                 jail.create_file(super::constants::CONFIG_WORKSPACE_FILENAME, "node: 123")?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -457,7 +507,7 @@ projects:
   foo: packages/foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -478,7 +528,7 @@ projects:
   foo: packages/foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -499,7 +549,7 @@ projects:
   foo: packages/foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -521,7 +571,7 @@ projects:
   foo: packages/foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -540,7 +590,7 @@ projects:
   foo: packages/foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -548,9 +598,9 @@ projects:
 
         #[test]
         fn inherits_from_env_var() {
-            env::set_var("MOON_NODE_VERSION", "4.5.6");
-
             figment::Jail::expect_with(|jail| {
+                jail.set_env("MOON_NODE_VERSION", "4.5.6");
+
                 jail.create_file(
                     super::constants::CONFIG_WORKSPACE_FILENAME,
                     r#"
@@ -560,20 +610,16 @@ projects: {}
 "#,
                 )?;
 
-                let config = super::load_jailed_config()?;
+                let config = super::load_jailed_config(jail.directory())?;
 
                 assert_eq!(config.node.version, String::from("4.5.6"));
 
                 Ok(())
             });
-
-            env::remove_var("MOON_NODE_VERSION");
         }
     }
 
     mod npm {
-        use super::*;
-
         #[test]
         #[should_panic(
             expected = "Invalid field <id>node.npm</id>: Expected struct NpmConfig type, received string \"foo\"."
@@ -588,7 +634,7 @@ node:
     npm: foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -612,7 +658,7 @@ projects:
 "#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -620,9 +666,9 @@ projects:
 
         #[test]
         fn inherits_from_env_var() {
-            env::set_var("MOON_NPM_VERSION", "4.5.6");
-
             figment::Jail::expect_with(|jail| {
+                jail.set_env("MOON_NPM_VERSION", "4.5.6");
+
                 jail.create_file(
                     super::constants::CONFIG_WORKSPACE_FILENAME,
                     r#"
@@ -634,19 +680,16 @@ projects: {}
 "#,
                 )?;
 
-                let config = super::load_jailed_config()?;
+                let config = super::load_jailed_config(jail.directory())?;
 
                 assert_eq!(config.node.npm.version, String::from("4.5.6"));
 
                 Ok(())
             });
-
-            env::remove_var("MOON_NPM_VERSION");
         }
     }
 
     mod pnpm {
-        use super::*;
 
         #[test]
         #[should_panic(
@@ -662,7 +705,7 @@ node:
     pnpm: foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -685,7 +728,7 @@ projects:
   foo: packages/foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -693,9 +736,9 @@ projects:
 
         #[test]
         fn inherits_from_env_var() {
-            env::set_var("MOON_PNPM_VERSION", "4.5.6");
-
             figment::Jail::expect_with(|jail| {
+                jail.set_env("MOON_PNPM_VERSION", "4.5.6");
+
                 jail.create_file(
                     super::constants::CONFIG_WORKSPACE_FILENAME,
                     r#"
@@ -708,19 +751,16 @@ projects: {}
 "#,
                 )?;
 
-                let config = super::load_jailed_config()?;
+                let config = super::load_jailed_config(jail.directory())?;
 
                 assert_eq!(config.node.pnpm.unwrap().version, String::from("4.5.6"));
 
                 Ok(())
             });
-
-            env::remove_var("MOON_PNPM_VERSION");
         }
     }
 
     mod yarn {
-        use super::*;
 
         #[test]
         #[should_panic(
@@ -736,7 +776,7 @@ node:
     yarn: foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -759,7 +799,7 @@ projects:
   foo: packages/foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -767,9 +807,9 @@ projects:
 
         #[test]
         fn inherits_from_env_var() {
-            env::set_var("MOON_YARN_VERSION", "4.5.6");
-
             figment::Jail::expect_with(|jail| {
+                jail.set_env("MOON_YARN_VERSION", "4.5.6");
+
                 jail.create_file(
                     super::constants::CONFIG_WORKSPACE_FILENAME,
                     r#"
@@ -782,14 +822,12 @@ projects: {}
 "#,
                 )?;
 
-                let config = super::load_jailed_config()?;
+                let config = super::load_jailed_config(jail.directory())?;
 
                 assert_eq!(config.node.yarn.unwrap().version, String::from("4.5.6"));
 
                 Ok(())
             });
-
-            env::remove_var("MOON_YARN_VERSION");
         }
     }
 
@@ -808,7 +846,7 @@ projects: {}
                     "projects: apps/*",
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -828,7 +866,7 @@ projects:
   foo: packages/foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -848,7 +886,7 @@ projects:
   foo: packages/foo"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -865,7 +903,7 @@ projects:
   foo: ./packages/foo"#,
                 )?;
 
-                let config = super::load_jailed_config()?;
+                let config = super::load_jailed_config(jail.directory())?;
 
                 assert_eq!(
                     config.projects,
@@ -890,7 +928,7 @@ projects:
     - 'packages/*'"#,
                 )?;
 
-                let config = super::load_jailed_config()?;
+                let config = super::load_jailed_config(jail.directory())?;
 
                 assert_eq!(
                     config.projects,
@@ -923,7 +961,7 @@ vcs:
     manager: svn"#,
                 )?;
 
-                let config = super::load_jailed_config()?;
+                let config = super::load_jailed_config(jail.directory())?;
 
                 assert_eq!(
                     config,
@@ -958,7 +996,7 @@ projects: {}
 vcs: 123"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -978,7 +1016,7 @@ vcs:
     manager: unknown"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });
@@ -998,7 +1036,7 @@ vcs:
     defaultBranch: 123"#,
                 )?;
 
-                super::load_jailed_config()?;
+                super::load_jailed_config(jail.directory())?;
 
                 Ok(())
             });